# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API of biophonie (https://secret-garden-77375.herokuapp.com/). Files are located in \"assets/\"",
        "title": "Swagger for biophonie-api",
        "termsOfService": "TODO",
        "contact": {
            "name": "TODO",
            "url": "TODO",
            "email": "TODO"
        },
        "license": {
            "name": "GPL-3.0 license",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/geopoint/closest/to/{latitude}/{longitude}": {
            "get": {
                "description": "retrieve the closest geopoint to another geopoint, excluding other geopoints",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geopoint"
                ],
                "summary": "get the closest geopoint",
                "parameters": [
                    {
                        "type": "number",
                        "description": "latitude",
                        "name": "latitude",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "longitude",
                        "name": "longitude",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "srid to project",
                        "name": "srid",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "multi",
                        "description": "optional ids to exclude from search",
                        "name": "not[]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/geopoint.ClosestGeoId"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    }
                }
            }
        },
        "/geopoint/{id}": {
            "get": {
                "description": "retrieve the geopoint in the database using its id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geopoint"
                ],
                "summary": "get a geopoint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "geopoint id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/geopoint.GeoPoint"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    }
                }
            }
        },
        "/geopoint/{id}/picture": {
            "get": {
                "description": "located in assets/",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geopoint"
                ],
                "summary": "get the picture filename",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "geopoint id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    }
                }
            }
        },
        "/geopoint/{id}/sound": {
            "get": {
                "description": "located in assets/",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geopoint"
                ],
                "summary": "get the sound filename",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "geopoint id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    }
                }
            }
        },
...
        "/restricted/geopoint": {
            "post": {
                "description": "create the geopoint in the database and save the sound and picture file (see testgeopoint dir)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geopoint"
                ],
                "summary": "create a geopoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "geopoint infos in a utf-8 json file",
                        "name": "geopoint",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "geopoint sound",
                        "name": "sound",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "geopoint picture",
                        "name": "picture",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/geopoint.GeoPoint"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    }
                }
            }
        },
        "/restricted/geopoint/{id}/enable": {
            "patch": {
                "description": "make the geopoint available",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Geopoint"
                ],
                "summary": "make the geopoint available",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "geopoint id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    }
                }
            }
        },
...
        "/restricted/user/{id}": {
            "patch": {
                "description": "make a user admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "make a user admin",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user is now admin",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "create a user in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "create user",
                "parameters": [
                    {
                        "description": "Add user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.AddUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.ErrMsg": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "malformed request"
                }
            }
        },
        "geopoint.ClosestGeoId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 18
                }
            }
        },
        "geopoint.GeoPoint": {
            "type": "object",
            "properties": {
                "amplitudes": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "example": [
                        0,
                        1,
                        2,
                        3,
                        45,
                        3,
                        2,
                        1
                    ]
                },
                "available": {
                    "type": "boolean",
                    "example": true
                },
                "createdOn": {
                    "type": "string",
                    "example": "2022-05-26T11:17:35.079344Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "picture": {
                    "type": "string",
                    "example": "https://example.com/picture-1.jpg"
                },
                "sound": {
                    "type": "string",
                    "example": "https://example.com/sound-2.wav"
                },
                "title": {
                    "type": "string",
                    "example": "Forêt à l'aube"
                },
                "userId": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}